renamed drawer index file and story for drawer component
