+++ packages/frontend-app/src/app/components/Drawer/Drawer.scss
