+++ packages/frontend-app/src/app/components/Modal/index.tsx
@@ -8,7 +8,7 @@ import "./index.scss";
 import useMediaQuery, { screens } from "../../hooks/useMediaQuery";
 import useSsr from "../../hooks/useSsr";
 import { Button } from "../Button/Button";
-import { Drawer } from "../Drawer";
+import { Drawer } from "../Drawer/Drawer";
 import { CancelIcon } from "../icons";
 
 import type { IProps as IButtonProps } from "../Button/Button";
